+++ src/id_translation/transform/__init__.py
@@ -1,0 +1,7 @@
+"""User-defined transformations of IDs and translations."""
+
+from ._impl.bitmask import BitmaskTransformer
+
+__all__ = [
+    "BitmaskTransformer",
+]