+++ tests/mapping/test_mapper_hierarchy.py
@@ -45,11 +45,12 @@ def run(
     use_short_circuit,
 ):
     values = NUMBER_OF_LEGS.copy()
+    score = HeuristicScore(
+        score_function=lambda v, c, cxt: [float(c == NUMBER_OF_LEGS[v]) for c in c],  # type:ignore[var-annotated]
+        heuristics=[ShortCircuit.dogs_have_4_legs] if use_short_circuit else (),
+    )
     mapper = Mapper(
-        HeuristicScore(
-            score_function=lambda v, c, cxt: [float(c == NUMBER_OF_LEGS[v]) for c in c],  # type:ignore[var-annotated]
-            heuristics=[ShortCircuit.dogs_have_4_legs] if use_short_circuit else (),
-        ),
+        score,
         overrides=StaticOverride.nobody_gets_any_legs if use_static_override else None,
         filter_functions=[(FilterFunction.nobody_has_4_legs, {})] if use_filter else (),
         cardinality=cardinality,