+++ tests/test_factory.py
@@ -4,7 +4,7 @@ import pytest
 
 from id_translation import Translator
 from id_translation.exceptions import ConfigurationError
-from id_translation.factory import TranslatorFactory, default_fetcher_factory
+from id_translation.factory import default_fetcher_factory
 from id_translation.fetching import AbstractFetcher, MemoryFetcher
 from id_translation.types import IdType, SourceType
 
@@ -29,19 +29,6 @@ def test_default_fetcher_factory(
 ) -> None:
     fetcher: AbstractFetcher[str, int] = default_fetcher_factory(clazz, dict(data={}))
     assert isinstance(fetcher, expected_type)
-
-
-@pytest.mark.parametrize(
-    "clazz",
-    [
-        None,
-        "id_translation.Translator",
-        "id_translation._translator.Translator",
-        Translator,
-    ],
-)
-def test_resolve_class(clazz):
-    assert TranslatorFactory.resolve_class(clazz) == Translator
 
 
 def test_missing_config():